Combine fetching and importing drug tariff
