+++ deploy/fetch_drug_tariff.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 . /webapps/openprescribing/.venv/bin/activate
-python /webapps/openprescribing/openprescribing/manage.py fetch_drug_tariff --settings=openprescribing.settings.production
+python /webapps/openprescribing/openprescribing/manage.py fetch_and_import_drug_tariff --settings=openprescribing.settings.production
